# Faster slotted array