--- a/src/Paprika.Benchmarks/Paprika.Benchmarks.csproj
+++ b/src/Paprika.Benchmarks/Paprika.Benchmarks.csproj
@@ -17,8 +17,4 @@
       <ProjectReference Include="..\Paprika\Paprika.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="BenchmarkDotNet.Artifacts\" />
-    </ItemGroup>
-
 </Project>