This PR slightly amends the search in `SlottedArray` making it a bit faster (~5%) for colliding hashes. This is based on not creating the `NibblePath` when preamble is different and comparison for the key can default much faster. The comparison should happen only when the hash is the same and the preamble (length + odd) is the same.

### Benchmarks

| Method          | Mean     | Error    | StdDev   | Code Size |
|---------------- |---------:|---------:|---------:|----------:|
| Hash_collisions (before) | 19.92 us | 0.389 us | 0.701 us |   3,222 B |
| Hash_collisions (after) | 18.35 us | 0.366 us | 0.697 us |   3,359 B |